Much more user-friendly approach to dealing with the alphabet
The tool needs the know all the possible characters of the
alphabet very early in order to build a Trie data structure.
That's why having the TRY directive in the affix file with
all the possible characters of the used alphabet is the
ideal scenario.

However in reality some dictionaries simply don't use the
TRY derective at all. Some of the other dictionaries do
have this directive, but it misses a character or two. There
are also some dictionaries with, for example, unwanted stray
Latin characters used as a part of Cyrillic words.

This patch first tries to use the characters from the TRY
directive (assuming that it's even available). But if this
fails and unexpected characters are encountered, then the
tool aborts the processing with an error message, scans
the whole AFF and DIC files to extract the full alphabet
and restarts processing from scratch. This incurs a minor
performance penalty, so having a proper TRY directive in
the AFF file is still highly encouraged.