Fix freeze on encoding testenc20250326.aff+csv
The usage of a prefix may be caused by a suffix rule.
Example affix rules:
  PFX A Y 1
  PFX A 0 a .
  SFX B Y 1
  SFX B 0 b/A .
And a stem example:
  foo/B

Here we have the prefix rule A getting triggered by the suffix
flag B from the "foo/B" stem. The encoder didn't understand
that and was trying to unsuccessfully clear the non-existing
flag A from a stem candidate. Since the result of this
unsuccessful flag clearing entered the queue and was
processed again, we effectively got an endless loop.