Some of the affix rules don't include the condition field
Still process such files, but issue a warning. While at it,
also parse the tags fields of the affix rules. Even though
Hunspell doesn't treat the '#' character in any special way,
assume that it's a comment after the tags field. Otherwise
the tags handling is going to be messy, considering that a
lot of existing dictionaries de-facto follow this convention.